pyproject.toml
[tool.poetry]
name = "v2ray-update-sub"
version = "0.1.0"
description = ""
authors = ["Yuxin Yan <yuxin.yan.cs@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.26.1"
schedule = "^1.2.2"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
